+++ packages/lib/src/index.ts
@@ -7,7 +7,7 @@ import getTheme from './getTheme';
 export * from './helpers';
 export { validateChildren, isComponentType } from './validateChildren';
 export { createSyntheticEvent } from './createSyntheticEvent';
-export { type Mutable } from './types';
+export type { Mutable, RecursiveRecord } from './types';
 
 export { typeIs, createUniqueClassName, getNodeTextContent, getTheme, Theme };
 export type { DarkModeProps, LgIdProps };