+++ packages/menu/src/Menu/Menu.tsx
@@ -63,6 +63,7 @@ export const Menu = React.forwardRef<HTMLDivElement, MenuProps>(function Menu(
     trigger,
     portalClassName,
     portalContainer,
+    portalRef,
     scrollContainer,
     popoverZIndex,
     ...rest
@@ -158,6 +159,7 @@ export const Menu = React.forwardRef<HTMLDivElement, MenuProps>(function Menu(
           usePortal,
           portalClassName,
           portalContainer,
+          portalRef,
           scrollContainer,
         }
       : { spacing, usePortal }),
@@ -266,6 +268,12 @@ Menu.propTypes = {
   align: PropTypes.oneOf(Object.values(Align)),
   justify: PropTypes.oneOf(Object.values(Justify)),
   refEl: PropTypes.shape({
+    current:
+      typeof window !== 'undefined'
+        ? PropTypes.instanceOf(Element)
+        : PropTypes.any,
+  }),
+  portalRef: PropTypes.shape({
     current:
       typeof window !== 'undefined'
         ? PropTypes.instanceOf(Element)