+++ packages/tabs/src/utils/index.ts
@@ -1,1 +1,2 @@
+export { getEnabledIndices } from './getEnabledIndices';
 export { getTestUtils } from './getTestUtils/getTestUtils';